requirements.txt
flask>=3.0.0
mecab-python3>=1.0.9
unidic>=1.1.0
